# api/admin.openapi.yaml
openapi: 3.1.0
...
info:
  title: Latchway Admin API
  version: 1.0.0
  summary: Canonical control-plane API used by the dashboard and CLI.
  description: |
    Normative draft for administrative bootstrap, tenancy, immutable configuration,
    secrets, users, installations, usage, audit and self-tests. Cookie-authenticated
    mutations also require origin validation and `X-CSRF-Token`; scoped bearer API
    tokens are not subject to browser CSRF but receive the same capability checks.
    Browser sessions receive the CSRF value in both the login/bootstrap response
    header and a Secure, SameSite=Strict, non-HttpOnly `__Host-latchway_csrf`
    cookie so a same-origin console reload can recover without persisting it.
    Every successful or rejected mutation produces a redaction-safe audit event.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://gateway.example.test
security:
  - AdminCookie: []
  - AdminToken: []
...
paths:
...
  /admin/v1/admin-sessions:
    get:
      tags: [Administrator sessions]
      operationId: listAdminSessions
      summary: List redaction-safe administrator sessions in the active organization.
      description: |
        Requires `manage_owners`. Returns administrator identity and lifecycle
        metadata only. Token hints, hashes, CSRF material, network metadata, and
        revocation reasons are never returned. `current` is true only when the
        request itself authenticated with that cookie session; bearer-token
        requests report false for every item.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Bounded administrator-session metadata page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminSessionMetadataPage' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    AdminCookie:
      type: apiKey
      in: cookie
      name: __Host-latchway_admin
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-admin-api-token
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, minLength: 1, maxLength: 2048 }
    PageSize:
      name: page_size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  headers:
...
    RequestID:
      description: Server-selected request identifier.
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    Problem:
      description: RFC 9457 problem using a stable code from `error-codes.yaml`.
      headers:
        X-Latchway-Request-ID: { $ref: '#/components/headers/RequestID' }
        Retry-After: { schema: { type: string } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    AdminUserID:
      type: string
      pattern: '^adm_[A-Za-z0-9_-]{16,128}$'
    AdminSessionID:
      type: string
      pattern: '^asn_[A-Za-z0-9_-]{16,128}$'
...
    PageInfo:
      type: object
      additionalProperties: false
      required: [has_more]
      properties:
        has_more: { type: boolean }
        next_cursor: { type: string, minLength: 1, maxLength: 2048 }
...
    ValidationIssue:
      type: object
      additionalProperties: false
      required: [severity, code, path, message]
      properties:
        severity: { type: string, enum: [error, warning] }
        code: { type: string, pattern: '^[a-z][a-z0-9_]{0,127}$' }
        path: { type: string, maxLength: 1024 }
        message: { type: string, maxLength: 2048 }
...
    AdminSessionMetadata:
      type: object
      additionalProperties: false
      required: [id, administrator, created_at, last_seen_at, expires_at, status, current]
      properties:
        id: { $ref: '#/components/schemas/AdminSessionID' }
        administrator:
          type: object
          additionalProperties: false
          required: [id, email]
          properties:
            id: { $ref: '#/components/schemas/AdminUserID' }
            email: { type: string, format: email, maxLength: 320 }
        created_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        status: { type: string, enum: [active, expired, revoked] }
        current: { type: boolean }
    AdminSessionMetadataPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/AdminSessionMetadata' } }
        page: { $ref: '#/components/schemas/PageInfo' }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, code, request_id, retryable, documentation_url]
      properties:
        type: { type: string, format: uri }
        documentation_url:
          type: string
          format: uri
          pattern: '^https://docs\.latchway\.dev/errors/[a-z0-9-]+$'
        title: { type: string, minLength: 1, maxLength: 256 }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string, minLength: 1, maxLength: 2048 }
        instance: { type: string, format: uri-reference }
        code:
          $ref: './client.openapi.yaml#/components/schemas/Problem/properties/code'
        request_id: { type: string, minLength: 8, maxLength: 128 }
        retryable: { type: boolean }
        retry_after: { type: string, format: date-time }
        operation_id: { type: string, pattern: '^arq_[0-7][0-9A-HJKMNPQRSTVWXYZ]{25}$' }
        errors: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/ValidationIssue' } }
      allOf:
        - if:
            properties:
              code: { const: operation_indeterminate }
            required: [code]
          then:
            required: [operation_id]
          else:
            not:
              required: [operation_id]